graphserver_proxy_py/pyproject.toml
[project]
name = "graphserver-proxy-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph-sdk>=0.1.70",
    "python-dotenv>=1.1.0",
]
